src/main/java/com/findme/FindMeAI/requirements.txt
Flask
boto3
torch
opencv-python-headless
pillow
mysql-connector-python
python-dotenv